requirements_dev.txt
django_debug_toolbar==4.2.0
pytest==7.4
pytest-django==4.7.0
pre-commit==3.6.0
factory_boy==3.3.0
faker==22.0.0
